fix(stdint): make int64_t and uint64_t actually 64 bits wide
`lib/inc/stdint.h` defined them as `int` and `unsigned int`, so
`sizeof(uint64_t) == 4`. Nothing could warn about it: a cast to a type that
turns out to be narrower than advertised is perfectly legal, so code that cast
to `uint64_t` specifically to gain range gained nothing and compiled clean.

They are now `long long` and `unsigned long long`, and the widths of every
type in the header are pinned with the compile-time check the tree already
uses elsewhere, so the next such change fails the build. Reverting the two
typedefs now gives:

    lib/inc/stdint.h:94:14: error: size of array 'int64_width_check' is negative
    lib/inc/stdint.h:95:14: error: size of array 'uint64_width_check' is negative
    kernel/src/drivers/ata.c:139:14: error: size of array 'ata_identity_size_check' is negative

INT64_MIN, INT64_MAX and UINT64_MAX did not exist and now do.

The third error above is the consequence the issue did not know about.
`ata_identity_t` describes the 512-byte IDENTIFY response, and the read
derives its word count from the structure:

    inportsw(dev->io_reg.data, (uint16_t *)&dev->identity,
             (sizeof(ata_identity_t) / sizeof(uint16_t)));

Measured from the debug info of the compiled object, the structure was **508**
bytes and is now **512**, with `sectors_48` at offset 200 either way, which is
where word 100 belongs. So the driver was asking for 254 of the 256 words the
drive had to give and leaving two in its buffer. Nothing was misread — the
truncated tail is the `unused7` padding — but the transfer was short, and the
size is now asserted against the 512 the hardware sends.

The toolchain consequence the issue warned about does not arise. A clean build
of the whole tree succeeds and needs no 64-bit helpers: `nm -u` finds no
`__udivdi3`, `__divdi3`, `__moddi3` or shift helper anywhere in the kernel
archive, so nothing performs 64-bit division or a variable 64-bit shift.

Widening the type did change the argument size at four call sites that passed
one of these values to a 32-bit conversion, and none of them produced a
diagnostic, because the logging helpers carry no `format` attribute (#216).
They are narrowed explicitly now, each with a comment saying why: the kernel's
`vsprintf` has no length modifiers at all, so there is no way to print a
64-bit value. That is filed as #332, along with the live `%lld` instance at
`sync.c:68` it uncovered. The ATA offset path narrowing that this makes
visible is #333.

`t_stdint` covers what a compile-time check cannot: that the arithmetic
carries past 32 bits at run time, on the shapes the kernel depends on — the
`(uint64_t)value + 1` of the elf.c bounds checks and the
`(uint64_t)block_index * block_size` of the ext2 offsets — and that the
narrower types still have their own widths, since widening everything would
pass the 64-bit checks and break every on-disk and hardware layout.

Verified: 67 tests, 0 failures, 0 panics, and the same with
-DENABLE_SCHEDULER_FEEDBACK=ON, which is the only configuration that compiles
one of the four narrowed call sites. -DENABLE_EXT2_TRACE=ON builds clean.

+++ kernel/src/io/proc_system.c
@@ -160,7 +160,12 @@ int procs_module_init(void)
 /// @param buffer the buffer.
 /// @param bufsize the buffer size.
 /// @return the amount we wrote.
-static ssize_t procs_do_uptime(char *buffer, size_t bufsize) { return sprintf(buffer, "%d", timer_get_seconds()); }
+// The uptime is a uint64_t and this printf has no 64-bit conversion, so it
+// is narrowed here rather than read half-width by a %d (#270).
+static ssize_t procs_do_uptime(char *buffer, size_t bufsize)
+{
+    return sprintf(buffer, "%u", (uint32_t)timer_get_seconds());
+}
 
 /// @brief Write the version inside the buffer.
 /// @param buffer the buffer.
@@ -194,10 +199,10 @@ static int procs_do_mounts_cb(super_block_t *sb, void *ctx)
         return 1;
     }
 
-    const char *source = (sb->source[0] != '\0') ? sb->source : sb->name;
+    const char *source     = (sb->source[0] != '\0') ? sb->source : sb->name;
     const char *mountpoint = sb->path;
-    const char *fstype = (sb->type && sb->type->name) ? sb->type->name : "unknown";
-    const char *options = "rw";
+    const char *fstype     = (sb->type && sb->type->name) ? sb->type->name : "unknown";
+    const char *options    = "rw";
 
     int written = snprintf(
         mounts_ctx->buf + mounts_ctx->pos,
@@ -213,7 +218,7 @@ static int procs_do_mounts_cb(super_block_t *sb, void *ctx)
 
     if ((size_t)written >= mounts_ctx->bufsize - mounts_ctx->pos) {
         // Buffer filled, stop writing.
-        mounts_ctx->pos = mounts_ctx->bufsize - 1;
+        mounts_ctx->pos                  = mounts_ctx->bufsize - 1;
         mounts_ctx->buf[mounts_ctx->pos] = '\0';
         return 1;
     }
